BAH-4656 | Refactor. Extract out ObservationRenderer as separate widget
