+++ apps/clinical/public/locales/locale_en.json
@@ -317,7 +317,11 @@
   "RADIOLOGY_INVESTIGATION_NAME": "Procedure Name",
   "RADIOLOGY_VIEW_IMAGES": "View Images",
   "RADIOLOGY_VIEW_REPORT": "View Report",
-  "RADIOLOGY_VIEW_QC": "View QC",
+  "RADIOLOGY_VIEW_QC": "View Quality Assessment",
+  "ERROR_LOADING_OBSERVATIONS": "Error loading observations",
+  "NO_OBSERVATIONS_AVAILABLE": "No observations available",
+  "NO_REPORT_DATA": "No report data available",
+  "NO_QUALITY_ASSESSMENT_DATA": "No quality assessment data available",
   "REACTIONS": "Reaction(s)",
   "REVOKED_STATUS": "Cancelled",
   "SELECT_ENCOUNTER_TYPE": "Select encounter type",
