+++ apps/clinical/public/locales/locale_es.json
@@ -317,7 +317,11 @@
   "RADIOLOGY_INVESTIGATION_NAME": "Nombre del Procedimiento",
   "RADIOLOGY_VIEW_IMAGES": "Ver Imágenes",
   "RADIOLOGY_VIEW_REPORT": "Ver Informe",
-  "RADIOLOGY_VIEW_QC": "Ver QC",
+  "RADIOLOGY_VIEW_QC": "Ver Evaluación de Calidad",
+  "ERROR_LOADING_OBSERVATIONS": "Error al cargar observaciones",
+  "NO_OBSERVATIONS_AVAILABLE": "No hay observaciones disponibles",
+  "NO_REPORT_DATA": "No hay datos de informe disponibles",
+  "NO_QUALITY_ASSESSMENT_DATA": "No hay datos de evaluación de calidad disponibles",
   "REACTIONS": "Reacción(es)",
   "REVOKED_STATUS": "Cancelado",
   "SELECT_ENCOUNTER_TYPE": "Seleccionar tipo de encuentro",
