+++ packages/bahmni-widgets/src/observationsRenderer/index.ts
@@ -1,0 +1,2 @@
+export { ObservationsRenderer, default } from './ObservationsRenderer';
+export type { ObservationsRendererProps } from './ObservationsRenderer';
